+++ src/canvas/canvas.class.ts
@@ -602,6 +602,9 @@ export class SelectableCanvas<
     this.renderCanvas(this.contextContainer, this._objectsToRender);
   }
 
+  /**
+   * text selection is rendered by the active text instance during the rendering cycle
+   */
   renderTopLayer(ctx: CanvasRenderingContext2D): void {
     ctx.save();
     if (this.isDrawingMode && this._isCurrentlyDrawing) {
@@ -619,6 +622,7 @@ export class SelectableCanvas<
   /**
    * Method to render only the top canvas.
    * Also used to render the group selection box.
+   * Does not render text selection.
    */
   renderTop() {
     const ctx = this.contextTop;